+++ wiki-explorer/src/main/resources/static/db/schema.sql
@@ -3,19 +3,19 @@ use WikiExplorer;
 
 
 create table if not exists Query(
-	id LONG PRIMARY KEY AUTO_INCREMENT,
+	id INT PRIMARY KEY AUTO_INCREMENT,
 	title VARCHAR(100),
 	time LONG
 );
 
 create table if not exists Word(
-	id LONG PRIMARY KEY AUTO_INCREMENT,
+	id INT PRIMARY KEY AUTO_INCREMENT,
 	value VARCHAR(100)
 );
 
 create table if not exists Occurences(
-	id LONG PRIMARY KEY AUTO_INCREMENT,
-	query_id LONG NOT NULL,
+	id INT PRIMARY KEY AUTO_INCREMENT,
+	query_id INT NOT NULL,
 	FOREIGN KEY (query_id) REFERENCES Query(id),
 	word_id LONG NOT NULL,
 	FOREIGN KEY (word_id) REFERENCES Word(id),
